git-svn-id: svn://123.57.173.36/meijia-parent@678 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/model/user/Users.java
@@ -51,7 +51,7 @@ public class Users {
 
 	private Date birthDay;
 
-	private Short degreeId;
+	private Long degreeId;
 
 	private String major;   
 
@@ -276,13 +276,11 @@ public class Users {
 		this.birthDay = birthDay;
 	}
 
-
-	public Short getDegreeId() {
+	public Long getDegreeId() {
 		return degreeId;
 	}
 
-
-	public void setDegreeId(Short degreeId) {
+	public void setDegreeId(Long degreeId) {
 		this.degreeId = degreeId;
 	}
